+++ discord_announcer/__init__.py
@@ -1,4 +1,6 @@
 """Initialize the app"""
 
-__version__ = "0.0.17"
+__version__ = "0.0.18"
 __title__ = "discord_announcer"
+__url__ = "https://github.com/fthomas-de/discord_announcer"
+__compat_date__ = "2026-05-19"