+++ discord_announcer/provider.py
@@ -1,7 +1,17 @@
-from esi.clients import EsiClientProvider
+from esi.openapi_clients import ESIClientProvider
+
+from discord_announcer import __compat_date__, __title__, __url__, __version__
 
 
-class DA_Provider(EsiClientProvider):
+class DA_Provider(ESIClientProvider):
     pass
 
-esi = DA_Provider()
+esi = DA_Provider(
+    compatibility_date=__compat_date__,
+    ua_appname=__title__,
+    ua_version=__version__,
+    ua_url=__url__,
+    operations=[
+        "GetCorporationsCorporationIdWalletsDivisionTransactions",
+    ],
+)