+++ discord_announcer/views.py
@@ -1,30 +1,45 @@
 """App Views"""
 
-from datetime import datetime, timedelta
-from discord_announcer.app_settings import TIME_DELTA
-from discord_announcer.selects import get_transactions_for_timeframe
-from discord_announcer.utilities import format_sales
+# Django
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
 from django.core.handlers.wsgi import WSGIRequest
 from django.http import HttpResponse
-from django.shortcuts import render
-from discord_announcer.discord_bot import send_message_to_discord
-from dateutil.tz import tzlocal
+from django.shortcuts import redirect, render
+from django.utils.translation import gettext_lazy as _
 
-# @login_required
-# @permission_required("discord_announcer.basic_access")
-# def index(request: WSGIRequest) -> HttpResponse:
-#     """
-#     Index view
-#     :param request:
-#     :return:
-#     """
+from discord_announcer import __version__
+from discord_announcer.forms import AnnouncerConfigFormSet
+from discord_announcer.models import AnnouncerConfig
 
-    # context = {"text": "Hello, World!"}
 
-    # sales = get_transactions_for_timeframe(datetime.now(tz=tzlocal())-timedelta(days=14, hours=TIME_DELTA))
-    # sales_formatted = format_sales(sales)
-    # print(sales_formatted)
-    # send_message_to_discord(messages=sales_formatted, channel_id=CHANNEL_ID, hours=TIME_DELTA)
+@login_required
+@permission_required("discord_announcer.basic_access")
+def index(request: WSGIRequest) -> HttpResponse:
+    """
+    Manage the announcer configurations (corporation, division, channel,
+    lookback window) that the scheduled task reads.
+    :param request:
+    :return:
+    """
 
-    # return render(request, "discord_announcer/index.html", context)
+    if request.method == "POST":
+        formset = AnnouncerConfigFormSet(
+            request.POST, queryset=AnnouncerConfig.objects.all()
+        )
+
+        if formset.is_valid():
+            formset.save()
+            messages.success(request, _("Configuration saved."))
+
+            return redirect("discord_announcer:index")
+    else:
+        formset = AnnouncerConfigFormSet(queryset=AnnouncerConfig.objects.all())
+
+    context = {
+        "title": _("Discord Announcer"),
+        "version": __version__,
+        "formset": formset,
+    }
+
+    return render(request, "discord_announcer/index.html", context)